+++ src/CryptoCurrency.Core/Exchange/Model/TradeItem.cs
@@ -16,11 +16,11 @@ namespace CryptoCurrency.Core.Exchange.Model
 
         public Epoch Created { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
 
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
 
         public CurrencyCodeEnum FeeCurrencyCode { get; set; }
