+++ src/CryptoCurrency.Core/Extensions/MarketAggregateExtensions.cs
@@ -13,13 +13,13 @@ namespace CryptoCurrency.Core.Extensions
             switch(candleType)
             {
                 case CandleTypeEnum.Open:
-                    return aggregates.Select(a => a.Open).ToArray();
+                    return aggregates.Select(a => (double)a.Open).ToArray();
                 case CandleTypeEnum.High:
-                    return aggregates.Select(a => a.High).ToArray();
+                    return aggregates.Select(a => (double)a.High).ToArray();
                 case CandleTypeEnum.Low:
-                    return aggregates.Select(a => a.Low).ToArray();
+                    return aggregates.Select(a => (double)a.Low).ToArray();
                 case CandleTypeEnum.Close:
-                    return aggregates.Select(a => a.Close).ToArray();
+                    return aggregates.Select(a => (double)a.Close).ToArray();
                 default:
                     throw new ArgumentException($"Unhandled candle type: {candleType}", "candleType");
             }
