+++ src/CryptoCurrency.Core/Market/Ohlc.cs
@@ -13,15 +13,15 @@ namespace CryptoCurrency.Core.Market
 
         public Epoch Epoch { get; set; }
 
-        public double Open { get; set; }
+        public decimal Open { get; set; }
 
         public Epoch OpenEpoch { get; set; }
 
-        public double High { get; set; }
+        public decimal High { get; set; }
 
-        public double Low { get; set; }
+        public decimal Low { get; set; }
 
-        public double Close { get; set; }
+        public decimal Close { get; set; }
 
         public Epoch CloseEpoch { get; set; }
     }
