+++ src/CryptoCurrency.ExchangeClient/Binance/Model/BinanceNewOrder.cs
@@ -7,10 +7,10 @@ namespace CryptoCurrency.ExchangeClient.Binance.Model
     public class BinanceNewOrderFill
     {
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("qty")]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         [JsonProperty("commission")]
         public string Commission { get; set; }
@@ -34,16 +34,16 @@ namespace CryptoCurrency.ExchangeClient.Binance.Model
         public long TransactTime { get; set; }
 
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("origQty")]
-        public double OriginalQuantity { get; set; }
+        public decimal OriginalQuantity { get; set; }
         
         [JsonProperty("executedQty")]
-        public double ExecutedQuantity { get; set; }
+        public decimal ExecutedQuantity { get; set; }
 
         [JsonProperty("cummulativeQuoteQty")]
-        public double CumulativeQuoteQuantity { get; set; }
+        public decimal CumulativeQuoteQuantity { get; set; }
 
         [JsonProperty("status")]
         public string Status { get; set; }
