+++ src/CryptoCurrency.ExchangeClient/Binance/Model/BinanceTradeItem.cs
@@ -14,13 +14,13 @@ namespace CryptoCurrency.ExchangeClient.Binance.Model
         public int OrderId { get; set; }
         
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("qty")]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         [JsonProperty("commission")]
-        public double Commission { get; set; }
+        public decimal Commission { get; set; }
 
         [JsonProperty("commissionAsset")]
         public string CommissionAsset { get; set; }
