+++ src/CryptoCurrency.ExchangeClient/CoinbasePro/Model/CoinbaseProTick.cs
@@ -7,16 +7,16 @@ namespace CryptoCurrency.ExchangeClient.CoinbasePro.Model
     public class CoinbaseProTick
     {
         [JsonProperty(PropertyName = "price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty(PropertyName = "bid")]
-        public double Bid { get; set; }
+        public decimal Bid { get; set; }
 
         [JsonProperty(PropertyName = "ask")]
-        public double Ask { get; set; }
+        public decimal Ask { get; set; }
 
         [JsonProperty(PropertyName = "volume")]
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
 
         [JsonProperty(PropertyName = "time")]
         public DateTime Time { get; set; }
