+++ src/CryptoCurrency.ExchangeClient/Kraken/Model/KrakenTick.cs
@@ -7,13 +7,13 @@ namespace CryptoCurrency.ExchangeClient.Kraken.Model
     public class KrakenPairTick
     {
         [JsonProperty(PropertyName = "a")]
-        public ICollection<double> Ask { get; set; }
+        public ICollection<decimal> Ask { get; set; }
 
         [JsonProperty(PropertyName = "b")]
-        public ICollection<double> Bid { get; set; }
+        public ICollection<decimal> Bid { get; set; }
 
         [JsonProperty(PropertyName = "c")]
-        public ICollection<double> Last { get; set; }
+        public ICollection<decimal> Last { get; set; }
     }
     
     public class KrakenTick : Dictionary<string, KrakenPairTick>
